+++ iddb-web/src/main/java/iddb/web/security/service/local/DbUserServiceImpl.java
@@ -166,7 +166,7 @@ public class DbUserServiceImpl extends CommonUserService {
 		
 		Player player = getSubjectPlayer(server);
 		
-		if (player != null && player.getLevel() != null && player.getLevel() >= level) {
+		if (player != null && player.getLevel() != null && player.getLevel().intValue() >= level.intValue()) {
 			return true;
 		}
 		return false;
